fix(controlplane): guard worker counter lookups against a missing counter
